Add case-insensitive support for string_table. More work is needed to enable it where appropriate.

+++ server/asClass.h
@@ -217,7 +217,7 @@ public:
 	{
 		if (getClassInternal(name))
 			return false;
-		mClasses[name] = a;
+		mClasses[static_cast<std::size_t>(name)] = a;
 		return true;
 	}
